// web/live-sports/src/views/message/utils/types.ts
export enum MessageType {
  text = 0,
  image = 1,
  voice = 2,
  video = 3,
  music = 4,
  news = 5,
}
// ...
export enum MessageCardType {
  SYSTEM = 'system',
  MESSAGE = 'message',
}
export type MessageCardData = {
  id: string;
  anchorId: string;
  sendUserId: string;
  name: string;
  cover: string;
  message: string;
  readDot: number;
  createTime: string;
  createTimeText: string;
  type: MessageCardType;
  msgType: MessageType;
};